results of numerous battery things
The results of the different charge rates seem to do nothing. The size of the battery, given an even mix of profiles and the same amount of solarpanels, is slightly noticeble at first glance. The supply is higher when the battery size increases and the avg SoC seems to stay similar (which means higher battery capacities have a definitive benefit)